Start/Stop Optstrings and Placeholder Method
Added a placeholder method to the Cli class for adding the stop code to
stop the daemon. Also added the start/stop/watch optstrings that add
the ability to start the service as a daemon, stop the daemon, and the
watch code which will allow a user to achieve the current start
functionality.